test: move mock fixtures to conftest.py to eliminate code duplication
