+++ tests/test_button.py
@@ -1,6 +1,5 @@
 import pytest
-from unittest.mock import AsyncMock, MagicMock
-from homeassistant.config_entries import ConfigEntry
+from unittest.mock import MagicMock
 
 from custom_components.frank_energie.const import (
     DATA_USER,
@@ -12,22 +11,6 @@ from custom_components.frank_energie.button import (
     FrankEnergieDisableSmartFeedInButton,
     FrankEnergieDisableSmartHvacButton,
 )
-
-
-@pytest.fixture
-def mock_coordinator():
-    coordinator = MagicMock()
-    coordinator.data = {}
-    coordinator.api = AsyncMock()
-    coordinator.async_request_refresh = AsyncMock()
-    return coordinator
-
-
-@pytest.fixture
-def mock_config_entry():
-    entry = MagicMock(spec=ConfigEntry)
-    entry.entry_id = "test_entry_id"
-    return entry
 
 
 @pytest.mark.asyncio
