+++ tests/test_switch.py
@@ -1,6 +1,5 @@
 import pytest
-from unittest.mock import AsyncMock, MagicMock
-from homeassistant.config_entries import ConfigEntry
+from unittest.mock import MagicMock
 
 from custom_components.frank_energie.const import (
     DOMAIN,
@@ -9,22 +8,6 @@ from custom_components.frank_energie.const import (
 from custom_components.frank_energie.switch import (
     FrankEnergieEnodeSmartChargingSwitch,
 )
-
-
-@pytest.fixture
-def mock_coordinator():
-    coordinator = MagicMock()
-    coordinator.data = {}
-    coordinator.api = AsyncMock()
-    coordinator.async_request_refresh = AsyncMock()
-    return coordinator
-
-
-@pytest.fixture
-def mock_config_entry():
-    entry = MagicMock(spec=ConfigEntry)
-    entry.entry_id = "test_entry_id"
-    return entry
 
 
 def test_enode_smart_charging_switch_properties(mock_coordinator, mock_config_entry):
